The implementation exercise of the teacher tried.

+++ RestWithASPNETUdemy/RestWithASPNETUdemy/Model/Book.cs
@@ -7,14 +7,8 @@ namespace RestWithASPNETUdemy.Model
     public class Book
     {
 
-//    id` INT(10) AUTO_INCREMENT PRIMARY KEY,
-//   `author` longtext,
-//   `launch_date` datetime(6) NOT NULL,
-//   `price` decimal(65,2) NOT NULL,
-//   `title` longtext
-
     [Column("id")]
-    public int Id {get; set;}
+    public long Id {get; set;}
 
     [Column("author")]
     public string Author {get; set;}
